[mapleall]Internal commit msg:     [codecheck 16]     temporary fix DTS2023030800677

+++ src/mapleall/maple_be/include/cg/cgfunc.h
@@ -1129,7 +1129,7 @@ class CGFunc {
     }
     auto it = func.GetLastFreqMap().find(stmt.GetStmtID());
     if (it != func.GetLastFreqMap().end()) {
-      frequency = it->second;
+      frequency = static_cast<uint32>(it->second);
     }
   }
