+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_memlayout.cpp
@@ -97,7 +97,7 @@ uint32 AArch64MemLayout::ComputeStackSpaceRequirementForCall(StmtNode &stmt,  in
       }
     }
     CCLocInfo ploc;
-    aggCopySize += parmLocator.LocateNextParm(*ty, ploc);
+    aggCopySize += static_cast<int32>(parmLocator.LocateNextParm(*ty, ploc));
     if (ploc.reg0 != 0) {
       continue;  /* passed in register, so no effect on actual area */
     }
